requirements.txt
python-binance>=1.0.19
pandas>=2.0
numpy>=1.24
requests>=2.28
